requirements.txt
# Core Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# QR Code and Barcode Decoding
pyzbar>=0.1.9

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0

# Data Handling
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Deep Learning (Optional - for advanced detection)
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Progress Bars
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
